Change sidebar to use data table Fixes #392
