+++ apps/web/src/app/admin/scenarios/client-section.tsx
@@ -1,16 +1,19 @@
 "use client";
 
 import { Scenario } from "@workspace/validators";
-import { ScenarioTable } from "./scenario-table";
+import { ScenarioTable } from "../../../components/scenario-table";
+import { useAdminColumns } from "./use-admin-columns";
 
 interface ClientSectionProperties {
   scenarios: Scenario[];
 }
 
 export default function ClientSection({ scenarios }: ClientSectionProperties) {
+  const columns = useAdminColumns();
+
   return (
     <div aria-label="Scenarios table container">
-      <ScenarioTable data={scenarios} />
+      <ScenarioTable data={scenarios} columns={columns} />
     </div>
   );
 }
